refactor: rename Generic to Items
