+++ packages/renderer/src/lib/help/HelpActions.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 import { Items } from '/@/lib/help/HelpItems';
 
-import HelpActionsGeneric from './HelpActionsGeneric.svelte';
+import HelpActionsItems from './HelpActionsItems.svelte';
 </script>
 
-<HelpActionsGeneric items={Items}/>
+<HelpActionsItems items={Items}/>
