+++ packages/renderer/src/lib/help/HelpActionsItems.svelte
